Move autocomplete to another package
