+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/code/injection/PlatformCodeEditorComponent.kt
@@ -1,6 +1,6 @@
 package org.hyperskill.app.android.code.injection
 
-import org.hyperskill.app.android.code.data.AutocompleteContainer
+import org.hyperskill.app.android.code.presentation.autocomplete.AutocompleteContainer
 import org.hyperskill.app.android.code.highlight.prettify.PrettifyParser
 import org.hyperskill.app.android.code.view.widget.CodeAnalyzer
