+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/code/presentation/autocomplete/AutocompleteState.kt
@@ -1,3 +1,3 @@
-package org.hyperskill.app.android.code.data
+package org.hyperskill.app.android.code.presentation.autocomplete
 
 data class AutocompleteState(val prefix: String, val words: List<String>)
