+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/code/view/adapter/CodeToolbarAdapter.kt
@@ -11,7 +11,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import org.hyperskill.app.android.R
-import org.hyperskill.app.android.code.data.AutocompleteState
+import org.hyperskill.app.android.code.presentation.autocomplete.AutocompleteState
 import org.hyperskill.app.android.code.presentation.model.symbolsForLanguage
 import ru.nobird.android.view.base.ui.extension.resolveColorAttribute
 import kotlin.math.abs
